Add support for cursors with highlight
The genhighlight.py script creates PNG files for a derived cursor
set which includes a filled, transparent circle (by default yellow)
around the hotspot of a cursor.  This makes the cursor better visible
when sharing or recording the screen.

The information needed is derived from the generated *.in and *.png
files and nothing else.  Any change to the graphics should be
automatically used after the PNG files for the normal cursors are
re-generated.

The script uses the pypng module which must be installed.

There are a few configuration parameters in the script:
- the color of the highlight circle
- the size of the highlight circle (as a fraction of the cursor size)