Change openhand symlink to hand1
Currently openhand is symlinked to left_ptr. This displays the
left_ptr cursor when interacting with a CSS cursor property set to
the grab value. The issue can be reproduced by visiting MDN's
page[0] for the cursor property.

Using GNOME's Adwaita cursor theme as a reference, the correct
cursor should look like an open hand. Changing the symlink for
openhand to hand1 displays the correct cursor.

[0]: https://developer.mozilla.org/en-US/docs/Web/CSS/cursor